+++ src/filesystem/fs.h
@@ -14,22 +14,22 @@
 
 struct fs_node;
 
-typedef uint32 (*read_type_t)(struct fs_node*,uint32,uint32,uint8*);
-typedef uint32 (*write_type_t)(struct fs_node*,uint32,uint32,uint8*);
+typedef size_t (*read_type_t)(struct fs_node*,size_t,size_t,uint8*);
+typedef size_t (*write_type_t)(struct fs_node*,size_t,size_t,uint8*);
 typedef void (*open_type_t)(struct fs_node*);
 typedef void (*close_type_t)(struct fs_node*);
-typedef struct dirent * (*readdir_type_t)(struct fs_node*,uint32);
+typedef struct dirent * (*readdir_type_t)(struct fs_node*,size_t);
 typedef struct fs_node * (*finddir_type_t)(struct fs_node*,char *name);
 
 typedef struct fs_node {
     char name[128];     
-    uint32 mask;        
-    uint32 uid;         
-    uint32 gid;         
-    uint32 flags;       
-    uint32 inode;       
-    uint32 length;      
-    uint32 impl;        
+    size_t mask;        
+    size_t uid;         
+    size_t gid;         
+    size_t flags;       
+    size_t inode;       
+    size_t length;      
+    size_t impl;        
     read_type_t read;
     write_type_t write;
     open_type_t open;
@@ -41,16 +41,16 @@ typedef struct fs_node {
 
 struct dirent {
     char name[128]; 
-    uint32 ino;     
+    size_t ino;     
 };
 
 extern fs_node_t *fs_root; 
 
-uint32 read_fs(fs_node_t *node, uint32 offset, uint32 size, uint8 *buffer);
-uint32 write_fs(fs_node_t *node, uint32 offset, uint32 size, uint8 *buffer);
+size_t read_fs(fs_node_t *node, size_t offset, size_t size, uint8 *buffer);
+size_t write_fs(fs_node_t *node, size_t offset, size_t size, uint8 *buffer);
 void open_fs(fs_node_t *node, uint8 read, uint8 write);
 void close_fs(fs_node_t *node);
-struct dirent *readdir_fs(fs_node_t *node, uint32 index);
+struct dirent *readdir_fs(fs_node_t *node, size_t index);
 fs_node_t *finddir_fs(fs_node_t *node, char *name);
 
 #endif
